WIP: Remove Content Tree and Course Selector under /features
